requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
amqp==5.2.0
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cryptography==42.0.5
decorator==5.1.1
Deprecated==1.2.14
execnet==2.1.1
fabric==3.2.2
fake==0.8
Faker==24.3.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
flower==2.0.1
gevent==24.2.1
geventhttpclient==2.2.1
greenlet==3.0.3
h11==0.14.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.3.5
locust==2.26.0
MarkupSafe==2.1.5
msgpack==1.0.8
mysql-connector-python==8.3.0
outcome==1.3.0.post0
packaging==24.0
paramiko==3.4.0
pluggy==1.4.0
progress==1.6
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pycparser==2.22
PyNaCl==1.5.0
PySocks==1.7.1
pytest==8.1.1
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.0.2
requests==2.31.0
roundrobin==0.0.4
selenium==4.23.0
setuptools==68.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.2
wheel==0.41.2
wrapt==1.16.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.3

playwright~=1.46.0